// src/glitchtip/generated/schema.d.ts
export interface components {
    schemas: {
// ...
        IssueActorSchema: {
            /**
             * Type
             * @enum {string}
             */
            type: "user" | "team";
            /** Id */
            id: string;
            /** Name */
            name: string;
            /** Username */
            username?: string | null;
            /** Email */
            email?: string | null;
            /** Slug */
            slug?: string | null;
        };
        /** IssueDetailSchema */
        IssueDetailSchema: {
            /** Id */
            id: string;
            /** Count */
            count: string;
            type: components["schemas"]["IssueEventTypeDisplay"];
            level: components["schemas"]["LogLevelDisplay"];
            status: components["schemas"]["EventStatusDisplay"];
            /** Metadata */
            metadata: {
                [key: string]: unknown;
            };
            project: components["schemas"]["ProjectReference"];
            /** Shortid */
            shortId: string;
            /** Numcomments */
            numComments: number;
            /**
             * Stats
             * @default {
             *       "24h": []
             *     }
             */
            stats: {
                [key: string]: number[][];
            } | null;
            /** Shareid */
            shareId?: number | null;
            /** Logger */
            logger?: string | null;
            /**
             * Permalink
             * @default Not implemented
             */
            permalink: string | null;
            /** Statusdetails */
            statusDetails?: {
                [key: string]: string;
            } | null;
            /** Subscriptiondetails */
            subscriptionDetails?: string | null;
            /**
             * Usercount
             * @default 0
             */
            userCount: number | null;
            /** Matchingeventid */
            matchingEventId?: string | null;
            firstRelease?: components["schemas"]["IssueReleaseSchema"] | null;
            lastRelease?: components["schemas"]["IssueReleaseSchema"] | null;
            /**
             * Firstseen
             * Format: date-time
             */
            firstSeen: string;
            /**
             * Lastseen
             * Format: date-time
             */
            lastSeen: string;
            assignedTo?: components["schemas"]["IssueActorSchema"] | null;
            /** Title */
            title: string;
            /** Culprit */
            culprit?: string | null;
            /** Userreportcount */
            userReportCount: number;
        };
// ...
        CommitSchema: {
            /** Id */
            id: string;
            /**
             * Message
             * @default
             */
            message: string | null;
            /** Datecreated */
            dateCreated?: string | null;
            /** Authorname */
            authorName?: string | null;
            /** Authoremail */
            authorEmail?: string | null;
        };
// ...
        IssueSchema: {
            /** Id */
            id: string;
            /** Count */
            count: string;
            type: components["schemas"]["IssueEventTypeDisplay"];
            level: components["schemas"]["LogLevelDisplay"];
            status: components["schemas"]["EventStatusDisplay"];
            /** Metadata */
            metadata: {
                [key: string]: unknown;
            };
            project: components["schemas"]["ProjectReference"];
            /** Shortid */
            shortId: string;
            /** Numcomments */
            numComments: number;
            /**
             * Stats
             * @default {
             *       "24h": []
             *     }
             */
            stats: {
                [key: string]: number[][];
            } | null;
            /** Shareid */
            shareId?: number | null;
            /** Logger */
            logger?: string | null;
            /**
             * Permalink
             * @default Not implemented
             */
            permalink: string | null;
            /** Statusdetails */
            statusDetails?: {
                [key: string]: string;
            } | null;
            /** Subscriptiondetails */
            subscriptionDetails?: string | null;
            /**
             * Usercount
             * @default 0
             */
            userCount: number | null;
            /** Matchingeventid */
            matchingEventId?: string | null;
            firstRelease?: components["schemas"]["IssueReleaseSchema"] | null;
            lastRelease?: components["schemas"]["IssueReleaseSchema"] | null;
            /**
             * Firstseen
             * Format: date-time
             */
            firstSeen: string;
            /**
             * Lastseen
             * Format: date-time
             */
            lastSeen: string;
            assignedTo?: components["schemas"]["IssueActorSchema"] | null;
            /** Title */
            title: string;
            /** Culprit */
            culprit?: string | null;
        };
        /** IssueTagSchema */
        IssueTagSchema: {
            /** Topvalues */
            topValues: components["schemas"]["IssueTagTopValue"][];
            /** Uniquevalues */
            uniqueValues: number;
            /** Key */
            key: string;
            /** Name */
            name: string;
            /** Totalvalues */
            totalValues: number;
        };
// ...
        IssueStatsResponse: {
            /** Id */
            id: string;
            /** Count */
            count: string;
            /** Usercount */
            userCount: number;
            /** Firstseen */
            firstSeen: string;
            /** Lastseen */
            lastSeen: string;
            /** Isunhandled */
            isUnhandled: boolean;
            stats: components["schemas"]["StatsDetailSchema"];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}